chore: move malleated transaction decoder to the encoding package
